iota_data_ingestion_core::history

Module manifest

Handle the manifest for historical checkpoint data.

MANIFEST File Disk Format ┌──────────────────────────────┐ │ magic<4 byte> │ ├──────────────────────────────┤ │ serialized manifest │ ├──────────────────────────────┤ │ sha3 <32 bytes> │ └──────────────────────────────┘

Structs

CheckpointUpdates

FileMetadata

ManifestV1

Enums

Manifest

Functions

create_file_metadata

create_file_metadata_from_bytes

finalize_manifest

Computes the SHA3 checksum of the Manifest and writes it to a byte vector.

read_manifest

Reads the manifest file from the store.

read_manifest_from_bytes

Reads the manifest file from the given byte vector and verifies the integrity of the file.

verify_historical_checkpoints_with_checksums

write_manifest

Writes the Manifest to the remote store.