iota_data_ingestion_core::history::reader

Function make_blob_iterator_for_range

pub fn make_blob_iterator_for_range(
    blob: Bytes,
    range: Range<CheckpointSequenceNumber>,
) -> IngestionResult<impl Iterator<Item = CheckpointData>>

Construct an iterator over a blob of checkpoint data.

The iterator filters checkpoints that belong to the specified range.

Errors

The function fails if the blob is corrupted and fails to decode.